feat(cli)!: Rename bins and env vars to cedar
