+++ docs/docs/cli-commands.md
@@ -1706,7 +1706,7 @@ export default async () => {
 You can pipe the script output to the formatter:
 
 ```bash
-yarn rw prisma db seed | yarn rw-log-formatter
+yarn rw prisma db seed | yarn cedar-log-formatter
 ```
 
 > Note: Just be sure to set `data` attribute, so the formatter recognizes the content.
@@ -2218,7 +2218,7 @@ For the full list of Server Configuration settings, see [this documentation](app
 If you want to format your log output, you can pipe the command to the Cedar LogFormatter:
 
 ```
-yarn rw serve api | yarn rw-log-formatter
+yarn rw serve api | yarn cedar-log-formatter
 ```
 
 ### serve web
@@ -2244,7 +2244,7 @@ This command serves the contents in `web/dist`. Use this command if you're debug
 If you want to format your log output, you can pipe the command to the Cedar LogFormatter:
 
 ```
-yarn rw serve web | yarn rw-log-formatter
+yarn rw serve web | yarn cedar-log-formatter
 ```
 
 ## upgrade
