+++ docs/versioned_docs/version-8.x/docker.md
@@ -273,7 +273,7 @@ ENV NODE_ENV=production
 # This is important if you intend to configure GraphQL to use Realtime.
 #
 # CMD [ "./api/dist/server.js" ]
-CMD [ "node_modules/.bin/rw-server", "api" ]
+CMD [ "node_modules/.bin/cedar-server", "api" ]
 ```
 
 :::important
