+++ packages/api-server/dist.test.ts
@@ -11,14 +11,15 @@ describe('dist', () => {
     expect(fs.existsSync(path.join(distPath, '__tests__'))).toEqual(false)
   })
 
-  it('ships three bins', () => {
+  it('ships the expected bins', () => {
     expect(packageConfig.bin).toMatchInlineSnapshot(`
       {
+        "cedar-api-server-watch": "./dist/cjs/watch.js",
+        "cedar-log-formatter": "./dist/cjs/logFormatter/bin.js",
+        "cedar-server": "./dist/cjs/bin.js",
         "cedarjs-api-server-watch": "./dist/watch.js",
         "cedarjs-log-formatter": "./dist/logFormatter/bin.js",
         "cedarjs-server": "./dist/bin.js",
-        "rw-api-server-watch": "./dist/cjs/watch.js",
-        "rw-log-formatter": "./dist/cjs/logFormatter/bin.js",
         "rw-server": "./dist/cjs/bin.js",
       }
     `)
