+++ packages/api-server/package.json
@@ -87,10 +87,12 @@
   "main": "./dist/createServer.js",
   "types": "./dist/createServer.d.ts",
   "bin": {
+    "cedar-api-server-watch": "./dist/cjs/watch.js",
+    "cedar-log-formatter": "./dist/cjs/logFormatter/bin.js",
+    "cedar-server": "./dist/cjs/bin.js",
     "cedarjs-api-server-watch": "./dist/watch.js",
     "cedarjs-log-formatter": "./dist/logFormatter/bin.js",
     "cedarjs-server": "./dist/bin.js",
-    "rw-api-server-watch": "./dist/cjs/watch.js",
     "rw-log-formatter": "./dist/cjs/logFormatter/bin.js",
     "rw-server": "./dist/cjs/bin.js"
   },
