+++ packages/cli/src/commands/serveWebHandler.ts
@@ -3,7 +3,7 @@ import execa from 'execa'
 import { getPaths } from '@cedarjs/project-config'
 
 export const webSsrServerHandler = async (rscEnabled?: boolean) => {
-  await execa('yarn', ['rw-serve-fe'], {
+  await execa('yarn', ['cedar-serve-fe'], {
     cwd: getPaths().web.base,
     stdio: 'inherit',
     env: rscEnabled
