+++ packages/core/package.json
@@ -11,8 +11,17 @@
   "type": "module",
   "bin": {
     "cedar": "./dist/bins/cedar.js",
+    "cedar-api-server-watch": "./dist/bins/cedar-api-server-watch.js",
+    "cedar-dev-fe": "./dist/bins/cedar-dev-fe.js",
     "cedar-gen": "./dist/bins/cedar-gen.js",
     "cedar-gen-watch": "./dist/bins/cedar-gen-watch.js",
+    "cedar-jobs": "./dist/bins/cedar-jobs.js",
+    "cedar-jobs-worker": "./dist/bins/cedar-jobs-worker.js",
+    "cedar-log-formatter": "./dist/bins/cedar-log-formatter.js",
+    "cedar-serve-api": "./dist/bins/cedar-serve-api.js",
+    "cedar-serve-fe": "./dist/bins/cedar-serve-fe.js",
+    "cedar-server": "./dist/bins/cedar-server.js",
+    "cedar-web-server": "./dist/bins/cedar-web-server.js",
     "cedarjs": "./dist/bins/cedar.js",
     "cedarjs-api-server-watch": "./dist/bins/cedarjs-api-server-watch.js",
     "cedarjs-serve-api": "./dist/bins/cedarjs-serve-api.js",
@@ -23,13 +32,9 @@
     "nodemon": "./dist/bins/nodemon.js",
     "redwood": "./dist/bins/redwood.js",
     "rw": "./dist/bins/redwood.js",
-    "rw-api-server-watch": "./dist/bins/rw-api-server-watch.js",
-    "rw-dev-fe": "./dist/bins/rw-dev-fe.js",
-    "rw-jobs": "./dist/bins/rw-jobs.js",
     "rw-jobs-worker": "./dist/bins/rw-jobs-worker.js",
     "rw-log-formatter": "./dist/bins/rw-log-formatter.js",
-    "rw-serve-api": "./dist/bins/rw-serve-api.js",
-    "rw-serve-fe": "./dist/bins/rw-serve-fe.js",
+    "rw-server": "./dist/bins/rw-server.js",
     "rw-web-server": "./dist/bins/rw-web-server.js",
     "vitest": "./dist/bins/vitest.js"
   },
