+++ packages/framework-tools/src/bins/cedar-fwtools-attw.ts
