+++ packages/jobs/src/consts.ts
@@ -14,7 +14,7 @@ export const DEFAULT_WORK_QUEUE = '*'
 export const DEFAULT_PRIORITY = 50
 export const DEFAULT_WAIT = 0
 export const DEFAULT_WAIT_UNTIL = null
-export const PROCESS_TITLE_PREFIX = 'rw-jobs-worker'
+export const PROCESS_TITLE_PREFIX = 'cedar-jobs-worker'
 // TODO: Change this to 'backgroundJob' in the next major release
 export const DEFAULT_MODEL_NAME = 'BackgroundJob'
