+++ packages/vite/attw.ts
@@ -7,12 +7,12 @@ interface Problem {
 }
 
 // Excluded entry points:
-// - ./bins/rw-vite-build.mjs: this is only used in the build handler
+// - ./bins/cedar-vite-build.mjs: this is only used in the build handler
 // - ./react-node-loader: used to run the Worker
 
 await $({
   nothrow: true,
-})`yarn attw -P --exclude-entrypoints ./bins/rw-vite-build.mjs -f json > .attw.json`
+})`yarn attw -P --exclude-entrypoints ./bins/cedar-vite-build.mjs -f json > .attw.json`
 const output = await $`cat .attw.json`
 await $`rm .attw.json`
