+++ packages/web-server/build.mts
@@ -6,22 +6,6 @@ import {
   defaultIgnorePatterns,
 } from '@cedarjs/framework-tools'
 
-// This package uses the name of the bin as `scriptName` for Yargs to keep things in sync.
-// There should only be one bin entry for this to work.
-// Otherwise we have to rethink the code.
-const { bin } = JSON.parse(fs.readFileSync('./package.json', 'utf8'))
-const bins = Object.keys(bin).length
-
-if (bins !== 1) {
-  console.error(
-    [
-      `Error: Expected exactly one bin entry; found ${bins}`,
-      './packages/web-server/src/bin.ts uses the bin entry as its scriptName',
-    ].join('\n'),
-  )
-  process.exit(1)
-}
-
 // Build the package
 await build({
   entryPointOptions: {
